requirements.txt
python-dotenv
ccxt
pybit
pandas
numpy
xgboost
scikit-learn
matplotlib
requests[socks]
autopep8
